fix cfg (toml does not support u128), pass eip1559_max_fee_per_gas_in_gwei to validator config

+++ ethexe/cli/src/params/ethereum.rs
@@ -59,12 +59,12 @@ pub struct EthereumParams {
     /// EIP-1559 max fee per gas in gwei for transaction fee estimation (for batch commitments).
     #[arg(long, alias = "eip1559-max-fee-per-gas-in-gwei")]
     #[serde(rename = "eip1559-max-fee-per-gas-in-gwei")]
-    pub eip1559_max_fee_per_gas_in_gwei: Option<u128>,
+    pub eip1559_max_fee_per_gas_in_gwei: Option<u64>,
 
     /// Blob gas multiplier.
     #[arg(long, alias = "blob-gas-multiplier")]
     #[serde(rename = "blob-gas-multiplier")]
-    pub blob_gas_multiplier: Option<u128>,
+    pub blob_gas_multiplier: Option<u64>,
 }
 
 impl EthereumParams {
@@ -82,11 +82,11 @@ impl EthereumParams {
         Ethereum::INCREASED_EIP1559_FEE_INCREASE_PERCENTAGE;
 
     /// Default EIP-1559 max fee per gas in gwei for transaction fee estimation (for batch commitments).
-    pub const DEFAULT_EIP1559_MAX_FEE_PER_GAS_IN_GWEI: u128 =
-        Ethereum::NO_EIP1559_MAX_FEE_PER_GAS_IN_GWEI;
+    pub const DEFAULT_EIP1559_MAX_FEE_PER_GAS_IN_GWEI: u64 =
+        Ethereum::NO_EIP1559_MAX_FEE_PER_GAS_IN_GWEI as u64;
 
     /// Default blob gas multiplier.
-    pub const DEFAULT_BLOB_GAS_MULTIPLIER: u128 = Ethereum::INCREASED_BLOB_GAS_MULTIPLIER;
+    pub const DEFAULT_BLOB_GAS_MULTIPLIER: u64 = Ethereum::INCREASED_BLOB_GAS_MULTIPLIER as u64;
 
     /// Converts Ethereum-facing CLI/TOML parameters into [`EthereumConfig`].
     ///
@@ -109,10 +109,12 @@ impl EthereumParams {
                 .unwrap_or(Self::DEFAULT_EIP1559_FEE_INCREASE_PERCENTAGE),
             eip1559_max_fee_per_gas_in_gwei: self
                 .eip1559_max_fee_per_gas_in_gwei
-                .unwrap_or(Self::DEFAULT_EIP1559_MAX_FEE_PER_GAS_IN_GWEI),
+                .unwrap_or(Self::DEFAULT_EIP1559_MAX_FEE_PER_GAS_IN_GWEI)
+                as u128,
             blob_gas_multiplier: self
                 .blob_gas_multiplier
-                .unwrap_or(Self::DEFAULT_BLOB_GAS_MULTIPLIER),
+                .unwrap_or(Self::DEFAULT_BLOB_GAS_MULTIPLIER)
+                as u128,
         })
     }
 }
